+++ src/app/FakeLib/EnvironmentHelper.fs
@@ -50,5 +50,14 @@ let ProgramFilesX86 =
     else
         environVar "ProgramFiles"
 
+let mutable TargetPlatformPrefix = @"C:\Windows\Microsoft.NET\Framework"
+
+/// Gets the local directory for the given target platform
+let getTargetPlatformDir platformVersion = 
+    if Directory.Exists(TargetPlatformPrefix + "64") then 
+        Path.Combine(TargetPlatformPrefix + "64",platformVersion) 
+    else 
+        Path.Combine(TargetPlatformPrefix,platformVersion)
+
 /// The path to the personal documents
 let documentsFolder = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments)